feat: step progress
