+++ src/components/StepProgressIndicator/messages.js
@@ -1,0 +1,11 @@
+import { defineMessages } from '@edx/frontend-platform/i18n';
+
+const messages = defineMessages({
+  progress: {
+    defaultMessage: '{done} of {needed} completed',
+    description: 'Progress indicator for assessments steps with multiple assessments',
+    id: 'frontend-app-ora.StepProgressIndicator.progress',
+  },
+});
+
+export default messages;
